Bug 557: config.h misses INSTALL_SUBDIR_* variables on first cmake run
The first time cmake runs in a new supertux build tree, config.h is generated
before the INSTALL_SUBDIR_{BIN,SHARE} variables are set and thus contains empty
values for these variables. If the executable is built right away and run, it
crashes because it cannot find the data files.

Resolves #557. Thanks to Matt McCutchen for this patch.

SVN-Revision: 6293