Bug 524: Fading TileMaps may change solidity of the TileMap.
When the alpha value of a layer is changed to > .75, the solid flag will be
set. When the alpha value of a layer is changed to < .25, the solid flag is
cleared. This prevents hiding background / foreground layers at the beginning
of the level and display them as a result to some action of the player.

Resolves #524.

SVN-Revision: 6342