object/unstable_tile.[ch]pp: Implement third state, "shake".
The order of the states is now:

  shake → dissolve → fall-down

Each state is optional. The tile will be made non-solid at the end of
"dissolve". If that action is missing in the sprite, the tile will be solid
when it falls down, making it possible to jump off of it.

Jumping off of a falling tile works better than I had expected. I've changed
the gravity modifier to 0.98 nonetheless, so that the tile falls slightly
slower than Tux does.

The "unstable_tile" and "castleblock" sprites now use the "shake" action and
not the "dissolve" action.

SVN-Revision: 6492