Cannon: Use the cannon's and badguy's bounding boxes to calculate the starting position.
The previous code assumed width of no more than 32 pixels to work correctly.
mathnerd314's change in revision 6530 widened that a bit, but could only be
considered a dirty hack. This change should work for arbitrary badguy widths,
moving narrow badguys (Short Fuse, for example) closer to the Cannon when shot
to the left.

SVN-Revision: 6542