Removed support for multiple Tilesets inside a level file, i.e. the (tilesets ...) tag
Neither the official nor addon levels make use of this feature, it
isn't support by any of the editors either. Addons that use custom
tilesets are just defining a new tileset, instead of mixing them in
the level file, so the ability to use custom tilesets inside a addon
is still available.

If there is any demand for this feature it should be brought back as
function inside the .strf files, not the .stl files, i.e. something
like:

(supertux-tiles
   (import-tileset
      (file "tilesetname.strf")
         (file "images/tiles/snow.stts")
         (start 0)
         (end   2153)
         (offset 0))

   ;; custom tiles go here

   ...))