Add test for collision.cpp
This test is for the intersects function that lies in supertux/collision.cpp
There are 5 testcases. 2 each for the conditions(seperated by ||) in each of
the if statements and one for the default case i.e. if it doesn't satisfy any
of the conditions. In default case it returns true else if it satisfies the
conditions in if statements it returns false.

Ref: #281