Snail: allow Tux to carry if flat
This feature was suggested by Karkus476. Because the snail is
the forest equivalent to the iceblock, it should be carreable too.
The snail class now inherits from the portable class, and implements
the methods grab and ungrab, and the snail states enum now also
has a STATE_GRABBED item.
Currently when "unthrown" the snails behaves exactly the same as when
you jump onto it, thus the snail flies a bit upwards first, it might
be better to have it move in a straight line when thrown.

closes #534