Show flying leaf animation for Walking Leaf and Leaf Shot
The flying leaf animation is now shown when the Walking Leaf and Leaf Shot
are squished. This animation was previously only shown for the Poison Ivy.
In order to limit code duplication, the spawn_explosion_sprites method was
created in the MovingSprite class. For the Walking Leaf, a new set of leaf
particles was created, based off of those for the Poison Ivy, but with
adjusted color. This fixes issue #644.