Renderer: Allow 4K viewports
This is necessary in order to fully be able to use the available resolution on
modern screens. An effect of this might be that some levels end abruptly at the
top or bottom edges of the screen, since those were designed for much lower
resolutions. However, that seems not any worse to me than showing a black edge.

Additionally, the code for badguys has been updated so they are aware of this
change,