Firefly: torch glow: smaller light
Use a smaller light sprite for the glow of the torch. To avoid centering
problems, bbox.get_middle() is used now rather than get_pos(). Additionally, a
new constant, TORCH_LIGHT_OFFSET, is introduced in order to position the light
sprite better, relative to the flame of the torch sprite.