scripting::TileMap: add set_solid() function
This allows changing the solidity of a tilemap on the fly without changing its
alpha value.

Not implementing a get_solid() function at this point, because we track to
values of solidity, the real (given) solidity and the effective solidity
(influenced by the tilemap's alpha value).

Reference: <https://forum.freegamedev.net/viewtopic.php?f=69&t=7849>