Revert "Renderer: Allow 4K viewports"
This reverts commit 1a54df0a0a3284cabd068356c83d96bbd867519f.

That patch broke most existing levels and leads to all kinds of
artifacts as the levels aren't even properly clipped.

Adding proper 4k support would either require fixing all those
levels or at the very least add proper clipping when the level
is smaller.

However 4k support itself might not be worth it, as
enemy spawn is tied to the screen size, thus it would lead to
enemies being spawned to far away from Tux and render some
gameplay situations impossible (e.g. enemy rolling down stairs
that Tux is trying to get up, on a large screen the enemy
would be on the ground before Tux reaches the stairs).

This could be worked around with aggro states for the enemies,
but that would require quite a bit of additional rework of all
the existing levels.