Make freetype required, libraqm optional
If libraqm was not found, then configure SDL_ttf --without-raqm and
don't link to libraqm.  This fixes a CMake error for linking to a
not-found library.  SuperTux built this way can't display Arabic but
can still display English.

Make Freetype REQUIRED, so not finding it is an error.  Update
INSTALL.md to list CMake 3.1, Freetype, and optional libraqm.