Remove this very broken feature (see reasons)
This removes the level count that would appear next to a world.  Why was this removed?  1) It looked plain ugly, but that's not really why I removed it.  2) It was a very buggy feature, that no one seemed to not be able to fix.  Not only was it that worldmaps would always be listed as NEW despite their actual status, but levelsets would tend to duplicate their level count and list more levels than they really had.  No one could find a fix to these bugs, and it is not a feature integral to what seems to be the current development plan (making the story mode good, seen as this feature affected only bonus levels)  However, the editor still gives the level counts.  This is not a bugfix, but rather a temporary disabling of this feature.