Resource packs, add-on dependencies, "TransferStatusList"-s (#2357)
Adds a new add-on variant - resource packs. Resource packs are different from other add-on types, since they are allowed to override game files. This allows modifying images, sounds, music and more. Some essential game folders have been restricted from being overridable by resource packs and the user directory.

Along with that, add-on dependencies are also included. When an add-on, which has dependencies, is installed, its dependencies would be installed as well, if they are not already and are available. When uninstalling an add-on, on which other add-ons depend, a warning message would be thrown to indicate that.

A new `TransferStatusList` class is implemented, which contains a vector of `TransferStatus` pointers. All `TranferStatus`-es are updated and tracked by their parent `TransferStatusList`. A `TranferStatusList` executes the provided callback (with `success` = `false`), if any `TransferStatus` has failed, and executes the callback (with `success` = `true`), if all `TransferStatus`-es have finished successfully. `DownloadDialog` now supports and operates using `TranferStatusList`-s.