Refactor WorldMap to support functional sectors (#2377)
Refactored `WorldMap` class, which supports functional sectors. Teleporting to a different sector in the same worldmap is now possible. This removes the need for using the unintuitive in most cases alternative, which is having multiple worldmaps.

`WorldMapSector` is the new class which handles sectors in a worldmap. It inherits `SectorBase`, which is a base for the `Sector` class that is used in levels, but only with its basic properties and functions included. This allows `WorldMapSector` to be parsed using `WorldMapSectorParser`, which inherits `SectorParser`.

Teleporting from one sector to another is possible both with scripting and using teleporters.

This pull request also features some general organizations/improvements to worldmap-related code:

* Editor worldmap object classes and main functional worldmap object classes have been merged into one for easier management.
* A fix for the scripting issue, previously attempted to be fixed in #2375, is incorporated in this PR, where the approach taken is to keep a reference to each `GameObject`'s parent `GameObjectManager`.

Fixes #1511.