Add a replace tool to the editor, clean up overlay_widget.cpp (#2539)
This tool was in the old Mono editor but has never been implemented in the in-game editor. It finds all instances of a tile in a tilemap and replaces them with the desired tile. This can be especially useful when a player creates their own tileset that doesn't have any autotiles for their custom tiles. With this feature, they can create a level using a theme that does have autotiles and then simply replace the tiles from the theme with corresponding tiles from their custom theme. It can also be used to change a level's theme in general.