Fix worldmap sector spawnpoint issues, testing from worldmap spawnpoints (#2547)
Fixes issues, related to loading a worldmap with a forced spawnpoint. It can now be set both with `set_initial_spawnpoint` on `WorldMap`, and by providing it to the initializer of the class. The forced spawnpoint overrides the saved Tux position state on the worldmap sector.

This PR also adds the ability to initialize a worldmap with a forced sector, allowing for teleporting on a different sector, other than "main", on a different worldmap.

Another addition is the "Test from here" option for worldmap spawnpoints. The worldmap will be tested from the current sector, Tux being on the position of the chosen spawnpoint to test from.