Fix translation issue with object display names
Previously, `ObjectFactory` saved copies of object display names on game startup, which would cause them to not be translatable, unless the game is restarted.
This is now fixed, since a factory function for getting the display name of an object is used instead.

The `m_other_display_names` member variable of `ObjectFactory` was removed as well. Its only use was to save the name of the node marker, since it is a part of the objects in the editor, while not being an actual object, but rather a tool.
It is planned to move the node marker to the toolbar in the future, so this variable would then be redundant.