Add a flip-action option for sprite actions (#2558)
This PR adds a "flip-action" option for sprite actions. It's a vertical version of the existing "mirror-action", which is used to remove the need for sprite images to be stored twice, once for each direction. This does the same for up and down.
As a result, UP images were removed for ispy, and replaced with flip-actions of down.