Move dart's sprite action initialisation to the constructor(s) (#2567)
Since the dart's direction is already known at the time of the Dart object being created, there's no reason the sprite action shouldn't be initialised in the constructor. This fixes an oversight from #2554 where the dart's position when spawning from a DartTrap is calculated from the sprite before its proper action is set, resulting in the maths not being correct for three of the four directions. I didn't notice this before since the dart is only spawned off by a few pixels, but it would be a bigger problem for custom dart sprites. Moving the sprite action to the constructor(s) fixes this, since the sprite action is already set properly at the time the dart's position is calculated when being shot out of a DartTrap.

EDIT: Alternatively, I could keep the sprite action initialisation in Dart::initialize() and set the action manually in DartTrap::fire() before Dart::initialize() is called. That would fix the problem too. Whichever you think is better.