Make sure `Rectf` width and height never go below 0
Previously, assertion in two of the initializers of `Rectf` was put in place, which checked if the rectangle has been initialized with negative width or height.

It was discovered that the assertion causes problems. If a hitbox has a width or height of 0, and is grown with a negative value as `border`. The newly initialized `Rectf` in the function would have negative width or height, so the game crashes on Debug builds, because of the assertions put in place.

The checks are now performed using `if` statements, and if the width or height go below 0, they are set to 0.
Additionally, these checks are also implemented in the width, height and size setters of `Rectf`.