Only show supported directions for each object that has Direction (#2572)
For most badguys, this is just `auto`, `left` and `right`. For some, `up` and `down` but not `auto`. For other objects, it varies - `Dart` has `auto`, `left`, `right`, `up`, `down`, `DartTrap` the same but without `auto`, `PushButton` has `up` and `down`, `Bumper` `left` and `right`. The `Switch` has `none`, `left`, `right`, `up` and `down` (I added `up` and `down` in this PR).

The bumper and the button used to have a `bool` instead of a `Direction`, I replaced those but made sure that levels saved with the old `bool` would load correctly and figure out the proper direction.

With switches, I replaced the three sprites with just one that supports directions. This does mean that old levels that used left or right switches will default to `switch.sprite` and show `Direction::NONE` switches, but that shouldn't matter much since those won't look out of place.

Also, code quality and guidelines in the files I touched.