Make new objects created in the editor face in their default direction rather than left. (#2582)
For some reason, the editor would put a (direction "left") entry to all newly created objects (this wouldn't save, but would mess with the object while being loaded after adding it). That includes objects that don't even support LEFT, like the PushButton, which is up and down only.

This fixes that.