Flush game objects, when finishing `WorldMapSector` construction
`WorldMapSector::finish_construction()`, just as with `Sector`, is called after the sector is done being parsed.
The function checks if required objects exist and if they don't, adds them.

The `WorldMapSector::finish_construction()` function was missing a `flush_game_objects()` call in the beginning, which caused the checks for required objects to always return `false`, hence creating duplicates of existing singleton objects.