Re-disable gravity for Sleeping Crystallo, fix missing sprite actions (#2588)
This adds back the missing jumping sprite actions for crystallo (I accidentally removed them in #2576 ).
It also re-disables gravity for Sleeping Crystallo, which I enabled in the same PR, but apparently, there is a legitimate use case and a reason for why it was disabled before, so I'm re-disabling it.
The gravity will also be disabled for Roof Crystallos created by waking roof-attached Sleeping Crystallos (until they drop).