Rewrite project license expression in AppData file (#2668)
The value for the "project_license" license tags in
supertuxkart2.appdata.xml had to be changed, since
"GPL-3.0+" is not a valid license expression, but
"GPL-3.0-or-later" is, according to the most recent
SPDX License List as of now
(version 3.22, 2023-10-05).

This is why, on Flathub, the SuperTux Flatpak app
shows its license as being "Proprietary"
when it's really not.
See: https://www.github.com/flathub/org.supertuxproject.SuperTux#20

This does not affect the build system or existing
game code.

[ci skip]