Uncomment `EditorToolbox` object display name warning
The warning was commented out, because prior to commit 545671a6711c0cb84552795da1a5a62b8f04e30a, the node marker was a part of the objects, so a warning would display every time it's hovered over.

Since the node marker is now displayed as a tool, the warning can be uncommented.