Profile manager, general profile info (#2666)
Adds a `ProfileManager`, which provides `Profile` objects, containing general data about a profile. The manager creates the profile, if it doesn't exist, so it always returns a valid `Profile` object.

Currently, the `Profile` object contains only the `name` property to indicate the displayed name for the profile on the profile menu.

This object is likely to be used for more properties in the future.