New Igel behavior + Corrupted Igel (#2648)
The current Igel behavior is rather annoying and not really intuitive. I propose changing it to the following:

When Igel is approached and they spot the player, they will quickly roll into a spiky ball and dash forward towards the player, unrolling after a couple seconds and continuing to walk as usual until they spot the player again. They will no longer turn around when shot with fire!

* Igel rolling animation [ci skip]

* Add sprites for corrupted variant of Igel [ci skip]

* Add missing roll sprites for Corrupted Igel [ci skip]

* Prepare code

* kjdshgoid

* keep rollin rollin rollin rollin

* igel roll stop

* a bunch of stuff

* easing

* corrutped igel

* bonk!

* fix

* Fix tabulation in corrupted_igel.sprite [ci skip]

* Fix tabulation in igel.sprite [ci skip]

* Fix include order [ci skip]

* Add missing const qualifiers pt. 1 [ci skip]

* Add missing const qualifiers pt. 2

* Add `get_` function prefix

* Fix implicit conversion

* Remove unused alias

---------

Co-authored-by: MatusGuy <85036874+MatusGuy@users.noreply.github.com>
Co-authored-by: Vankata453 <78196474+Vankata453@users.noreply.github.com>