Fix water not being electrocuted by thunderstorms and deprecate most tiles in the `others` menu. (#2828)
I added back the thunderstorm... thing... at the end of `tiles.strf` as it was removed some time ago.

I also deprecated a lot of tiles in the `others` tab because they're duplicates. They shouldn't be used and just having them in the others tab makes converted levels look fine even though they aren't.

For the tiles i didn't remove my reasons are:

- 1, 359: They're such classics. Make for nice test tiles. (we should add them into the misc tab or something)
- 2075, 2076, 2079,  2164, 2165: Special cases for tiles that need multiple tilemaps but used to use a single tilemap. Someone told me a reason they're not deprecated but I forgot. Would probably need discussion before deprecating.
- 3873-3877: Lightning water tiles. These are used to electrocute the water with thunder. Pretty simple. (The rain drop tiles aren't actually needed for rain particles, hence why i deprecated them. If someone can confirm otherwise i will undeprecate them.)
- 5658: This tile is missing from the `Forest Background` tab. Should probably ask Rusty where to put this as i couldn't find a good place
- 6539, 6540, 6545 Missing from the `Corrupted Forest` tab and missing from the autotile definitions leaving some corners looking a bit wonky. Would also need to ask Rusty where to put them. (Id need help adding them to the autotiles as i don't know what to do there.)

Any other tile was either a duplicate or shouldnt be used (like the end sequence tiles) except for an empty tile which was... well... emtpy.

---------

Co-authored-by: Vankata453 <78196474+Vankata453@users.noreply.github.com>