New Tarantula behavior + Spidermite Variant (#2820)
Tarantula, originally Spidermite, will be getting a behavior overhaul. Alongside it Tarantula will become its new default variant and is planned to make appearances in an upcoming world 3. The old Spidermite will get re-introduced as a Halloween variant of Tarantula for the Halloween Add-On, with a brand new design as well.

Their behavior will be a mixture of its old behavior and that of a Crusher!

Tarantula will reside idle on the ceiling, only charging downwards if Tux is close. Once they dropped down towards the floor, they will now bounce up and down within a set distance from the ground for as long as Tux remains close to it. Once Tux leaves, Tarantula will begin to slowly move up towards its original resting place and wait until Tux gets close again.