Editor: Adding layers from layers widget (#2884)
Adding layers to a sector now makes more sense, since it's done through a "+" button, found after the last layer on the layers widget. Pressing the button shows an object selection, just like in the toolbox widget, only containing the objects, formerly found in the "Sector" category for non-worldmap levels, and "Tilemap" and "Background" for worldmap levels. Pressing an object directly adds it to the sector.

To achieve this, most of the `EditorToolboxWidget` code, which managed tile/object selection, was moved to a different class. The resulting `EditorTilebox` class can also be used for features like tile backpacks in the future.

Additionally, this PR allows objects to be moved around right after the editor is started (without an objectgroup having to be active).

Closes #638.