Revert "Generate builds with `RelWithDebInfo` instead of `Release` (#2963)" (#3040)
This reverts commit a6d899c8548ae42330127831db2d738f23e588c2.

RelWithDebInfo include debug data separately from the executable, which needs
a debugger to be interpreted properly. The debug data does not affect stack
traces generated by the program itself, but increases the executable by more
than 100Mb, which is wasted if the user does not intent to run the exectuable
inside a debugger.

Co-authored-by: Semphris <semphris@protonmail.com>