Toggling level statistics (#3063)
The coins, badguys and secrets level statistics can now be toggled from a level's settings. By default, coins and secrets are enabled, but badguys are not (due to a recent discussion regarding the difficulty of making it possible to kill all badguys in a level; can be discussed further).

A level is shown as "perfect" on a worldmap if only the enabled statistics are maxed out.

All code which draws statistics has been updated to incorporate this change as well. Statistics which are turned off will not be shown alongside others.