Heavy coins can no longer be collected twice by falling on two hit blocks
This commit makes it so coins can only be collected once.

The exact issue leading up to this chance is that heavy coins used to be able to be collected twice, when hitting the top of two hit blocks at the same time.

This could be reproduced by having one coin explosion bonus block and another block (of any type), as well as a low ceiling. When hitting both blocks at the same time, some coins would fall in between of those 2 blocks, making them be collected twice, essentially duplicating them. The lower the ceiling is, the higher the chance of this occurring was.