[NOSQUASH] Editor: Replace the OKLab color picker by a 2D color picker (#2895)
* Editor: Add a 2D color picker

A two-dimensional image can enable a more convenient selection of colors than a one-dimensional slider.
Under the assumption that users typically want to select a color with the highest saturation,
a two-dimensional color picker where any RGB color with the highest saturation can be slected with one click could be suitable.
For an independent adjustment of lightness and hue, which is considered a good property of a color picker,
the two-dimensional color picker should use a perceptual color space such as OKLab.
As a side effect for choosing a perceptual color space,
a marker showing the lightness and hue of the current color can visualize how a change in RGB values affects the perceived color.

Changes:
* For the color selection in the editor, add the ItemColorPicker2D menu item,
  which displays an image and enables picking a color from it by clicking.
  The image contains the highest-saturation sRGB colors,
  so for the color picker we only need OKLab code to determine marker positions
  and no color clipping code since we can sample pixels from the image.
* Add a method drawing a hexagon to Canvas
* Add a helper method to ColorOKLCh which calculates a modified lightness
* Add a missing include guard and constness to the code for ColorOKLCh

* Editor: Remove the 1D OKLab color picker

There are reports that the one-dimensional OKLab color selection is unpopular and difficult to use,
so we can remove it to simplify the code and GUI.

This commit removes ItemColorChannelOKLab and most of the OKLab color space code.
Since the two-dimensional color picker uses some of the OKLab code,
it is not removed completely.