`GameSession`: Remote Restart, Restart from checkpoint and Abort support
Clients are now notified by the server of any of these events under the `CH_GAME_UPDATES` channel.

To support remote level restarting, `GameSession` now saves the levelfile content into `m_levelfile_content`, which is used by both servers and clients. Servers can use this variable to save the level with all respective object UIDs to a string and on restart, restore from there, just like the clients would do too. This also has the benefit of not having to read the level from the hard disk on every restart.