Item Pocket (#3070)
Original description by MatusGuy:
The Item Pocket allows you to save a powerup for later use. If you collect a flower while having a bonus greater than GROWUP_BONUS, that flower gets equipped and the old flower gets stored in the top left corner of the screen. Now, the player can press the new ITEM control (usually Select/Back or Left Shift) to use the stored flower by throwing it up and catching it.

This pull request replaces the powerup stacking feature because this new solution is much more balanced. It also limits the amount of concurrent players to 4.

* aeiou

* add item pocket input

* item pocket workie

* item pocket hud

* lifesaver texture

Co-authored-by: RustyBox@users.noreply.github.com

* item pocket bare minimum

* god fujjgking damngit

* THINGS

* pepsi pocket

Co-authored-by: RustyBox <RustyBox@users.noreply.github.com>

* this is for the good of the item economy

* more bugs fickesed

* powerup stacking is gone

* editor

* fix ci

* setting wip

* item pocket setting complete

* agjhag

* tdjyfjgfjgf

* bitch

* the anticheat mechanism

* omfg

* wip [ci skip]

* wip again [ci skip]

* script

* augh squirrel thing

* Apply suggestions from code review

Co-authored-by: Vankata453 <78196474+Vankata453@users.noreply.github.com>

* Fix being unable to spawn player [ci skip]

* Remove code duplication in `MultiplayerPlayerMenu`, do not remove players from status

[ci skip]

* Fix scripting documentation consistency with function

* wip script item pocket

* chagne docs

* fix duo player bug weird strange weird

* Scripting docs: Full stops

[ci skip]

* Fix `BonusType` enum descriptions for scripting documentation
[ci skip]

---------

Co-authored-by: MatusGuy <MatusGuy@users.noreply.github.com>
Co-authored-by: RustyBox <RustyBox@users.noreply.github.com>
Co-authored-by: Vankata453 <78196474+Vankata453@users.noreply.github.com>