Fix `PlayerStatusHUD` addition being pushed to `GameObjectManager` undo stack
`PlayerStatusHUD` is no longer added in editor levels. This fixes a bug where it would be pushed to the undo stack and opening a level in the editor without doing anything would still have it marked as having unsaved changes.

`PlayerStatusHUD` now also overrides `track_state` to indicate that it shouldn't be tracked, just in case.