Add "weakresourcepack" add-on type
This add-on type is also listed as "Resource Pack" to the user, however, it cannot override data files like a regular "resourcepack" could. This also means that more than 1 "weakresourcepack" can be active at the same time, as opposed to "resourcepack" being restricted to only 1.

Additionally adds an indicator in the add-on preview menu as to whether an add-on overrides data for more transparency.