Remove `RectfObjectOption`
`RectfObjectOption` used to save copies of the width and height of the provided rectangle, instead of a pointer to the rectangle itself. Additionally, it only parsed and wrote width and height, and was only used in `MovingObject`, rendering it useless against using two `FloatObjectOption`s for width and height.

This fixes resizing objects not being tracked in undo/redo.