Don't delete markers in editor after undo/redo
Deleting markers after undo/redo was done as a measure to remove invalid resize markers for objects, which might have been removed after undo/redo.

`ResizeMarker` now checks if its object is valid and if not, removes itself. This means deleting the markers in the editor after undo/redo as a measure shouldn't be needed anymore.