Fix weird "input spasm" bug
If a connected SDL GameController compatible device sends faulty axis
inputs, even if those inputs are below the deadzone, they will overwrite
the inputs sent by the keyboard, creating an odd "spasm effect" when
holding a directional input in the keyboard.

Here, in the GameControllerManager, I check the proposed input and the
current stick state, so that I can confirm that the GameController is
free to send inputs without interrupting anything else.