Use system-specific cursor for custom cursor
Currently, the game draws the cursor as it's own sprite. This is fine in most cases, however, this will cause input lag when the user moves their cursor. This is because the sprite cursor is drawn on every flip, which may lag behind.

The OS (which SDL interfaces) has it's own cursor api, and therefore is preferred, since that sets the display systems physical cursor instead of rendering our own.