opengl: Fancy effects setting
One of the largest complaints we receive is that this game's OpenGL
renderer is horrifically slow. This is undoubtly so, as we use a back
buffer (aka back_renderer) to perform the displacement effect (on item
tiles/ice/water) which essentially renders the game _twice_: one to
the backbuffer surface, then one to the actual GL framebuffer.

We will improve that later to use GL framebuffers, but for now, this
adds a "fancy effects" setting the user can toggle to turn off all
that nonsense.

To my knowledge, the wretched get_pixels functions should be OK, as
the OpenGL ES backend doesn't use a back renderer either. Has not been
tested though.

Signed-off-by: Swagtoy <me@ow.swag.toys>