Hide the properties sidebar option until it's ready
This feature isn't disabled! We want to ship the editor changes but I don't
believe the properties sidebar is ready yet, so for now, let's just set the
default option to false.

If you are reading this message and want to experiment with the properties
sidebar, edit $HOME/.local/share/supertux2/config, in the the editor car, add:

    (show_properties_sidebar #t)

And you can mess with it again.