Revert "Linked sprite support (#2753, #3418)"
There are numerous crashes and bugs that have been occuring due to this
commit. I think it just needs a little more time to bake in the oven. I'm sure
we could get the feature in the door by the time 0.7 releases, but it's simply
too risky to push out right before release since it involve too many critical
backend changes. For now, we revert :(

It's planned to put this back in after 0.7 releases though.

This reverts commit 7f1f74d5990c60b93846e2642b639fb260e67b34.