logger: --log-tinygettext to show tinygettext logs
With --debug, tinygettext tends to completely shit up the place. Despite that,
these logs are still useful, but they often get in the way of finding issues...

The approach here is pretty gross, but it's not like this setting is in the
config, so I just expose it lazily. It'll do for now.

Signed-off-by: Swagtoy <me@ow.swag.toys>