Fix the "slope ledges" problem
AKA a downwards slope against a wall.

This is a rather annoying bug that Frost and maybe some others i forgot
complained about. Essentially, for NE/NW slopes, you can cling onto the
sides. This however can be cheesed to avoid death and even skip entire sections,
and for a while it required level designers to essentially design against
it. That is no more (well, i think), as we do a somewhat nasty check if the
player is falling, as well as which side they are on, and then simply avoid the
rectangle constraint which is performed alongside the slope.

The "physics hint" variable basically implies that you can set that for any
object which may have any issues with collision that physics could help with. In
this case, we check the velocity (of the player) to make a conclusion on if we
should collide with a slope ledge

I'm not sure if this is the most humble solution, but it seems to work fine.

Signed-off-by: Swagtoy <me@ow.swag.toys>