control: Switch from Key to Scancode & fix infamous console key bug
Well, the latter is fixed inversely by this. Scancodes refer to it as grave, and
thus this translates to the backtick on a US QWERTY keyboard.

There's some degree of compatibility; I _tried_ to just convert them over, but
it seemed like SDL_GetScancodeFromKey kept returning 0. Maybe it was stubbed or
it's some weird bug. I don't know. I don't really care. People can just reset
their keybinds, it's really not the end of the world here since most people
leave them default, and at most, just change w,a,s,d or of course, that damned
console key on sane keyboard layouts.

And thus this also introduces the (version N) list locally to the control
list. It'll probably never be bumped, but at least it's there now. I also
removed the old (sdl2 #f) flag. Technically, this change invalidates that one
anyway since version doesn't exist and it too just resets the keys, and that
thing is old anyway, so why even keep it around?

Signed-off-by: Swagtoy <me@ow.swag.toys>